Place I think people will get stuck
Here I think what you're getting at is we should be worried about the variables that have the combo of high parametric uncertainty and high model sensitivity. But I think people are going to a little more of a lead in to the question or they won't get it. It seems kind of out of nowhere. This is my attempt at a prompt.